refactor(codex): split goal controls from app files
